fixing the last problema
* Fixed the problem of the tables when you go back
* Fixed the update problem

+++ Vista/Administrador/informacionPersonal.php
@@ -1,4 +1,5 @@
 <?php
+
 $idAdmin = $_SESSION['id'];
 
 if (isset($_POST['actualizarInfoAdministrador'])) {
@@ -9,34 +10,88 @@ if (isset($_POST['actualizarInfoAdministrador'])) {
     $oldUrl = $_POST['url_hidden'];
     $archivo = $_FILES['archivo']['name'];
 
-    if (isset($archivo) && $archivo != "") {
+    $cliente = new Cliente("", "", "", $correo);
+    $inventarista = new Inventarista("", "", "", $correo);
+    $administrador = new Administrador($idAdmin);
+    $administrador -> getInfoBasic();
+    
+    if ($administrador -> getCorreo() != $correo && ($inventarista->existeCorreo() || $cliente->existeCorreo() || $administrador -> existeNuevoCorreo($correo))) {
 
-        $archivo = date("Y_m_d_H_i_s_") . $archivo;
+        $msj = "El correo proporcionado ya se encuentra en uso.";
+        $class = "alert-danger";
 
-        $tipo = $_FILES['archivo']['type'];
-        $tamano = $_FILES['archivo']['size'];
-        $temp = $_FILES['archivo']['tmp_name'];
-        $url = 'static/img/Users/' . $archivo;
+    } else {
 
-        if (!((strpos($tipo, "gif") || strpos($tipo, "jpeg") || strpos($tipo, "jpg") || strpos($tipo, "png")) && ($tamano < 9000000))) {
+        if (isset($archivo) && $archivo != "") {
 
-            $class = "alert-danger";
-            $msj = "El tipo de archivo no es valido o el tamañano es demasiado grande";
-            $administrador = new Administrador($idAdmin);
-            $administrador->getInfoBasic();
+            $archivo = date("Y_m_d_H_i_s_") . $archivo;
+
+            $tipo = $_FILES['archivo']['type'];
+            $tamano = $_FILES['archivo']['size'];
+            $temp = $_FILES['archivo']['tmp_name'];
+            $url = 'static/img/Users/' . $archivo;
+
+            if (!((strpos($tipo, "gif") || strpos($tipo, "jpeg") || strpos($tipo, "jpg") || strpos($tipo, "png")) && ($tamano < 9000000))) {
+
+                $class = "alert-danger";
+                $msj = "El tipo de archivo no es valido o el tamañano es demasiado grande";
+                
+            } else {
+                if (move_uploaded_file($temp, $url)) {
+
+                    if (file_exists($oldUrl)) {
+                        unlink(trim($oldUrl));
+                    }
+
+                    $administrador = new Administrador($idAdmin, $nombre, $apellido, $correo, "", trim($url));
+
+                    $resInsert = $administrador->actualizarAdministrador();
+
+                    if ($resInsert == 1) {
+
+                        /**
+                         * Creo un objeto para retornar el dia y la hora
+                         */
+                        $date = new DateTime();
+
+                        if ($_SESSION['rol'] == 1) {
+                            /**
+                             * Creo el objeto de log
+                             */
+                            $logAdmin = new LogAdmin("", $date->format('Y-m-d H:i:s'), LogHActualizarAdministrador($idAdmin, $nombre, $apellido, $correo, $url), 22, getBrowser(), getOS(), $_SESSION['id']);
+                            /**
+                             * Inserto el registro del log
+                             */
+                            $logAdmin->insertar();
+
+                            /**
+                             * Log para el Inventarista
+                             */
+                        }
+
+                        $class = "alert-success";
+                        $msj = "El producto se ha guardado correctamente.";
+                    } else if ($resInsert == 0) {
+                        $class = "alert-warning";
+                        $msj = "No se ha modificado ningún valor.";
+                    }
+                    else {
+                        $class = "alert-danger";
+                        $msj = "Ocurrió algo inesperado";
+                    }
+                } else {
+                    $class = "alert-danger";
+                    $msj = "Ocurrió algo inesperado";
+                }
+            }
 
         } else {
-            if (move_uploaded_file($temp, $url)) {
 
-                if (file_exists($oldUrl)) {
-                    unlink(trim($oldUrl));
-                }
+            $administrador = new Administrador($idAdmin, $nombre, $apellido, $correo, "", trim($oldUrl));
 
-                $administrador = new Administrador($idAdmin, $nombre, $apellido, $correo, "", $url);
                 $resInsert = $administrador->actualizarAdministrador();
 
                 if ($resInsert == 1) {
-
                     /**
                      * Creo un objeto para retornar el dia y la hora
                      */
@@ -46,7 +101,7 @@ if (isset($_POST['actualizarInfoAdministrador'])) {
                         /**
                          * Creo el objeto de log
                          */
-                        $logAdmin = new LogAdmin("", $date->format('Y-m-d H:i:s'), LogHActualizarAdministrador($idAdmin, $nombre, $apellido, $correo, $url), 22, getBrowser(), getOS(), $_SESSION['id']);
+                        $logAdmin = new LogAdmin("", $date->format('Y-m-d H:i:s'), LogHActualizarAdministrador($idAdmin, $nombre, $apellido, $correo, $oldUrl), 22, getBrowser(), getOS(), $_SESSION['id']);
                         /**
                          * Inserto el registro del log
                          */
@@ -58,61 +113,26 @@ if (isset($_POST['actualizarInfoAdministrador'])) {
                     }
 
                     $class = "alert-success";
-                    $msj = "El producto se ha guardado correctamente.";
+                    $msj = "El producto se ha guardado correctamente";
                 } else if ($resInsert == 0) {
                     $class = "alert-warning";
                     $msj = "No se ha modificado ningún valor.";
-                } else {
+                }
+                else {
                     $class = "alert-danger";
                     $msj = "Ocurrió algo inesperado";
                 }
-            } else {
-                $class = "alert-danger";
-                $msj = "Ocurrió algo inesperado";
-            }
+            
         }
-        include "Vista/Main/error.php";
-    } else {
-
-        $administrador = new Administrador($idAdmin, $nombre, $apellido, $correo, "", $oldUrl);
-        $resInsert = $administrador->actualizarAdministrador();
-
-        if ($resInsert == 1) {
-            /**
-             * Creo un objeto para retornar el dia y la hora
-             */
-            $date = new DateTime();
-
-            if ($_SESSION['rol'] == 1) {
-                /**
-                 * Creo el objeto de log
-                 */
-                $logAdmin = new LogAdmin("", $date->format('Y-m-d H:i:s'), LogHActualizarAdministrador($idAdmin, $nombre, $apellido, $correo, $oldUrl), 22, getBrowser(), getOS(), $_SESSION['id']);
-                /**
-                 * Inserto el registro del log
-                 */
-                $logAdmin->insertar();
-
-                /**
-                 * Log para el Inventarista
-                 */
-            }
-
-            $class = "alert-success";
-            $msj = "El producto se ha guardado correctamente";
-        } else if ($resInsert == 0) {
-            $class = "alert-warning";
-            $msj = "No se ha modificado ningún valor.";
-        } else {
-            $class = "alert-danger";
-            $msj = "Ocurrió algo inesperado";
-        }
-
-        include "Vista/Main/error.php";
     }
+
+    include "Vista/Main/error.php";
+
 } else {
+
     $administrador = new Administrador($idAdmin);
     $administrador->getInfoBasic();
+    
 }
 
 
@@ -184,4 +204,3 @@ if (isset($_POST['actualizarInfoAdministrador'])) {
         </div>
     </div>
 </div>
-
